feat: interface cleanup
Various interfaces were comingling responsiblities, so these changes are intended to better organize and implement each. Any interactions directly with Arweave should be added to ArweaveDataProvider interface, and any interactions directly with Warp/Smartweave should extend the SmartweaveDataProvider interface. Additionally simplifies the implementation of ArweaveCompositeDataProviders to accept a single ArweaveDataProvider and SmartweaveDataProvider as are not yet in a position to need/require multiple providers and caches given the state of the broader Smartweave ecosystem.

Components and states are also updated with these additional changes, as well as some organization of various helper functions.