fix(PE-3913): update useTransactionDataHook to set params based on TransactionState
The way we had it, the TransactionState was updated based on URL params. This approach is very brittle, as a bad URL with the various type casting could easily break. Doing it this way, we can explicity set the TransactonState from anywhere in the app, and then update the URL accordingly. In a separate PR, we can add the logic that parses searchParams for transactionData values, and updates the TransactionState accordingly.